requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile
#
aiohttp==3.5.4            # via discord.py
aioprocessing==1.0.1
aiosqlite==0.10.0
aiostream==0.4.0
async-timeout==3.0.1      # via aiohttp
attrs==19.3.0             # via aiohttp
certifi==2019.9.11        # via requests
chardet==3.0.4            # via aiohttp, requests
ciso8601==2.1.2           # via tortoise-orm
discord.py==1.2.4
idna==2.8                 # via requests, yarl
multidict==4.5.2          # via aiohttp, yarl
praw==6.4.0
prawcore==1.0.1           # via praw
pypika==0.35.15           # via tortoise-orm
python-dotenv==0.10.3
requests==2.22.0          # via prawcore, update-checker
six==1.13.0               # via structlog, websocket-client
structlog==19.2.0
tortoise-orm==0.14.1
typing-extensions==3.7.4.1  # via tortoise-orm
update-checker==0.16      # via praw
urllib3==1.25.6           # via requests
websocket-client==0.56.0  # via praw
websockets==6.0           # via discord.py
yarl==1.3.0               # via aiohttp
